@@ -2,6 +2,7 @@
     "name": "sparkpost/php-sparkpost",
     "description": "SDK for interfacing with SparkPost APIs",
     "license": "Apache 2.0",
+    "version": "0.2.0",
     "authors": [
       {
         "name": "Message Systems, Inc."
@@ -14,17 +15,20 @@
       "test": "phpunit ./test/unit/"
     },
     "require": {
-    	"php": ">=5.3.0",
-        "guzzlehttp/guzzle": "3.8.1"
+      "php": ">=5.5.0",
+      "egeloen/http-adapter": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.3.*",
-	"satooshi/php-coveralls": "dev-master"
+      "phpunit/phpunit": "4.3.*",
+      "guzzlehttp/guzzle": "6.*",
+      "mockery/mockery": "^0.9.4",
+      "satooshi/php-coveralls": "dev-master"
     },
     "autoload": {
-    	"psr-4": {
-    		"SparkPost\\": "lib/SparkPost/",
-    		"SparkPost\\SendGridCompatibility\\": "lib/SendGridCompatibility/"
-    	}
+      "psr-4": {
+        "SparkPost\\": "lib/SparkPost/",
+        "SparkPost\\SendGridCompatibility\\": "lib/SendGridCompatibility/",
+        "SparkPost\\Test\\TestUtils\\": "test/unit/TestUtils/"
+      }
     }
 }
