@@ -3,27 +3,29 @@ namespace Examples\Transmisson;
 require_once (dirname(__FILE__).'/../bootstrap.php');
 
 //pull in API key config
-$configFile = file_get_contents(dirname(__FILE__) . "/../example-config.json");
+$configFile = file_get_contents(dirname(__FILE__) . '/../example-config.json');
 $config = json_decode($configFile, true);
 
 use SparkPost\SparkPost;
-use SparkPost\Transmission;
+use GuzzleHttp\Client;
+use Ivory\HttpAdapter\Guzzle6HttpAdapter;
 
-SparkPost::setConfig(array('key'=>$config['api-key']));
+$httpAdapter = new Guzzle6HttpAdapter(new Client());
+$sparky = new SparkPost($httpAdapter, ['key'=>$config['api-key']]);
 
 try {
 
-	$results = Transmission::send(array(
-		"campaign"=>"my-campaign",
-		"from"=>"From Envelope <from@example.com>",
-		"html"=>"<p>Hello World! Your name is: {{name}}</p>",
-		"text"=>"Hello World!",
-		"subject"=>"Example Email: {{name}}",
-		"recipientList"=>'Example List'
-	));
+  $results = $sparky->transmission->send([
+    'campaign'=>'my-campaign',
+    'from'=>'From Envelope <from@sparkpostbox.com>',
+    'html'=>'<p>Hello World! Your name is: {{name}}</p>',
+    'text'=>'Hello World!',
+    'subject'=>'Example Email: {{name}}',
+    'recipientList'=>'Example List'
+  ]);
 
-	echo 'Congrats you can use your SDK!';
+  echo 'Congrats you can use your SDK!';
 } catch (\Exception $exception) {
-	echo $exception->getMessage();
+  echo $exception->getMessage();
 }
 ?>
